requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.4
ecdsa==0.18.0
fastapi==0.101.0
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rq==1.15.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
supervisor==4.2.5
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websocket-client==1.6.2
websockets==11.0.3
